+++ routemaster/exit_conditions/__init__.py
@@ -1,9 +1,7 @@
 """Parsing and evaluation of exit condition programs."""
 
-from routemaster.exit_conditions.context import Context
 from routemaster.exit_conditions.program import ExitConditionProgram
 
 __all__ = (
-    'Context',
     'ExitConditionProgram',
 )
